+++ src/Components/Routes.tsx
@@ -11,6 +11,8 @@ import TokenDetail from "../Views/SecurityPanel/TokenDetail";
 import { TokenPermissions } from "reduct-js";
 import Replications from "../Views/Replications/Replications";
 import ReplicationDetail from "../Views/Replications/ReplicationDetail";
+import Lifecycles from "../Views/Lifecycles/Lifecycles";
+import LifecycleDetail from "../Views/Lifecycles/LifecycleDetail";
 import EntryDetail from "../Views/BucketPanel/EntryDetail";
 import QueryPanelPage from "../Views/QueryPanel/QueryPanelPage";
 
@@ -125,6 +127,24 @@ export function AppRoutes(props: Props): React.ReactElement {
         element={
           <PrivateRoute authorized={authorized} loading={loading}>
             <ReplicationDetail client={props.backendApi.client} {...props} />
+          </PrivateRoute>
+        }
+      />
+
+      <Route
+        path="/lifecycles"
+        element={
+          <PrivateRoute authorized={authorized} loading={loading}>
+            <Lifecycles client={props.backendApi.client} {...props} />
+          </PrivateRoute>
+        }
+      />
+
+      <Route
+        path="/lifecycles/:name"
+        element={
+          <PrivateRoute authorized={authorized} loading={loading}>
+            <LifecycleDetail client={props.backendApi.client} {...props} />
           </PrivateRoute>
         }
       />