+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add lifecycle management UI pages with list and detail views, [PR-206](https://github.com/reductstore/web-console/pull/206)
+
 ### Changed
 
 - Pin Docker GitHub Actions to latest stable commit SHAs in CI workflows, [PR-205](https://github.com/reductstore/web-console/pull/205)